Add comments to issues
In order to retrieve the comments field we have to explicitly ask for it.
This means we have to explicitly ask for ALL fields we are interested in. If
we are interested in a field that is not listed we have to add it.

We add the last comment as a property on the issue